@@ -7,5 +7,5 @@ import org.springframework.boot.runApplication
 class MyApplication
 
 fun main(args: Array<String>) {
-    runApplication<MyApplication>(*args)
+  runApplication<MyApplication>(*args)
 }
