@@ -1,5 +1,5 @@
 package com.rogervinas.stream.domain
 
 interface MyEventConsumer {
-    fun consume(event: MyEvent)
+  fun consume(event: MyEvent)
 }
