@@ -3,15 +3,14 @@ package com.rogervinas.stream.functional
 import com.rogervinas.stream.domain.MyEvent
 import com.rogervinas.stream.domain.MyEventConsumer
 import com.rogervinas.stream.shared.MyEventPayload
-import java.util.function.Consumer
 
-class MyStreamEventConsumer(private val consumer: MyEventConsumer) : Consumer<MyEventPayload> {
+class MyStreamEventConsumer(private val consumer: MyEventConsumer) : (MyEventPayload) -> Unit {
 
-    override fun accept(payload: MyEventPayload) {
-        consumer.consume(fromPayload(payload))
-    }
+  override fun invoke(payload: MyEventPayload) {
+    consumer.consume(fromPayload(payload))
+  }
 
-    private fun fromPayload(payload: MyEventPayload): MyEvent {
-        return MyEvent(payload.string)
-    }
+  private fun fromPayload(payload: MyEventPayload): MyEvent {
+    return MyEvent(payload.string)
+  }
 }
