@@ -1,9 +1,6 @@
 package com.rogervinas.stream.shared
 
-import com.fasterxml.jackson.annotation.JsonCreator
-import com.fasterxml.jackson.annotation.JsonProperty
-
-class MyEventPayload @JsonCreator constructor(
-        @JsonProperty("string") val string: String,
-        @JsonProperty("number") val number: Int
+data class MyEventPayload(
+  val string: String,
+  val number: Int
 )
