@@ -1,6 +1,10 @@
 package com.rogervinas.stream
 
-import com.nhaarman.mockito_kotlin.*
+import com.nhaarman.mockito_kotlin.any
+import com.nhaarman.mockito_kotlin.argumentCaptor
+import com.nhaarman.mockito_kotlin.doThrow
+import com.nhaarman.mockito_kotlin.timeout
+import com.nhaarman.mockito_kotlin.verify
 import com.rogervinas.stream.domain.MyEvent
 import com.rogervinas.stream.domain.MyEventConsumer
 import com.rogervinas.stream.domain.MyEventProducer
@@ -18,18 +22,20 @@ import org.springframework.boot.test.mock.mockito.MockBean
 import org.springframework.core.env.Environment
 import org.springframework.test.context.ActiveProfiles
 import java.time.Duration
-import java.util.*
+import java.util.UUID
 import java.util.function.Consumer
 
 @SpringBootTest(webEnvironment = DEFINED_PORT)
 @ActiveProfiles("docker-compose")
 class MyApplicationShould {
 
-  val TOPIC = "my.topic"
-  val TOPIC_DLQ = "my.topic.errors"
+  companion object {
+    private const val TOPIC = "my.topic"
+    private const val TOPIC_DLQ = "my.topic.errors"
 
-  val TEN_SECONDS = Duration.ofSeconds(10)
-  val FIVE = 5
+    private val TEN_SECONDS = Duration.ofSeconds(10)
+    private const val FIVE = 5
+  }
 
   @Autowired
   lateinit var env: Environment
